% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_dims_length_from_raw}
\alias{calc_dims_length_from_raw}
\title{Calculate dimension lengths from raw metadata}
\usage{
calc_dims_length_from_raw(metadata_raw)
}
\arguments{
\item{metadata_raw}{Raw metadata list from the INE API.}
}
\value{
A data.frame with dim_num and n columns.
}
\description{
Calculate dimension lengths from raw metadata
}
\keyword{internal}
