% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_dim_values}
\alias{extract_dim_values}
\title{Extract dimension values from raw metadata}
\usage{
extract_dim_values(metadata_raw)
}
\arguments{
\item{metadata_raw}{Raw metadata list from the INE API.}
}
\value{
A data.frame with dimension values.
}
\description{
Extract dimension values from raw metadata
}
\keyword{internal}
