% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_cached_data}
\alias{filter_cached_data}
\title{Filter cached data frame to match current dimension filters}
\usage{
filter_cached_data(data, current_filters, dim_values)
}
\arguments{
\item{data}{Cached data.frame}

\item{current_filters}{Normalized dimension filters from current request}

\item{dim_values}{Dimension values tibble from metadata (output of extract_dim_values)}
}
\value{
Filtered data.frame
}
\description{
Filter cached data frame to match current dimension filters
}
\keyword{internal}
