% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{finalize_chunk}
\alias{finalize_chunk}
\title{Finalize a chunk by validating and renaming from .part to .json}
\usage{
finalize_chunk(temp_path, final_path)
}
\arguments{
\item{temp_path}{Path to the temporary .part file}

\item{final_path}{Path to the final .json file}
}
\value{
TRUE on success, FALSE on failure
}
\description{
Finalize a chunk by validating and renaming from .part to .json
}
\keyword{internal}
