% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_ine_data}
\alias{process_ine_data}
\title{Process raw INE API responses into tidy dataframe}
\usage{
process_ine_data(raw_data)
}
\arguments{
\item{raw_data}{List containing parsed JSON responses and urls from fetch_data_raw()}
}
\value{
Tidy data.frame with INE data
}
\description{
Process raw INE API responses into tidy dataframe
}
\keyword{internal}
